Update requirements.rst (#18)
I have improved the English. The main problem was the ungrammatical use of commas. There is a difficulty in fixing those ungrammatical commas. The difficulty owes to the problem, or to a consequence of the problem. Here is what I mean. Ungrammatical commas leave the logical interrelations of clauses unclear. Does the clause after the comma mean to be the ground for the clause after the comma? Or are two thoughts being jammed into one? Thus the problem with fixing the commas is that I have had to _infer_ the intended logical relations.